+++ src/client/hud/layers/ControlPanel.ts
@@ -150,7 +150,11 @@ export class ControlPanel extends LitElement implements Controller {
       .outgoingAttacks()
       .map((a) => a.troops)
       .reduce((a, b) => a + b, 0);
-    this.troopRate = config.troopIncreaseRate(player) * 10;
+    this.troopRate =
+      config.troopIncreaseRate(
+        player,
+        this.game.numTilesWithFallout() / this.game.numLandTiles(),
+      ) * 10;
 
     const helpEnabled = new UserSettings().helpMessages();
 
@@ -342,7 +346,12 @@ export class ControlPanel extends LitElement implements Controller {
   private updateTroopGrowthSlope() {
     const player = this.game?.myPlayer();
     if (player === null) return;
-    const rate = this.game.config().troopIncreaseRate(player);
+    const rate = this.game
+      .config()
+      .troopIncreaseRate(
+        player,
+        this.game.numTilesWithFallout() / this.game.numLandTiles(),
+      );
     this._troopGrowthSlowing = rate < this._lastTroopIncreaseRate;
     this._lastTroopIncreaseRate = rate;
   }